Vp9decoding: Set the bo buffe object as cached to keep cache coherency
This is mainly for Cherryview. Although it has no LLC, it still can keep
the cache coherency between CPU and GPU after gfx buffer object is configured
as cached..
As BDW/Haswell has the LLC, it is already allocated as cached. And it is
harmless although it is called again.

Change-Id: I94cd49a17e819d1670dcc939c30d824c2aece83c
Signed-off-by: Zhao Yakui <yakui.zhao@intel.com>