vp9hdec: Cast ABS parameter to int to avoid C++11's `double abs(Integral)`
C++11 added an overload of abs with return type double for any
Integral parameter type in math.h.  This overload matches more
closely than the `int abs(int)` overload from stdlib.h if the
parameter type happens to be an integral type that isn't int (like
short for example, and ignoring long and long long here).

So to pick the `int abs(int)` overload we need to explicitly cast
ABS's parameter to int.

Signed-off-by: Scott D Phillips <scott.d.phillips@intel.com>